backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-multipart==0.0.20
Pillow==11.1.0
python-dotenv==1.0.1
pydantic-settings==2.7.1
aiofiles==24.1.0
